article-01-ai-native-pipeline/requirements.txt
anthropic>=0.40.0
boto3>=1.34.0
snowflake-connector-python>=3.6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
apache-airflow>=2.8.0
dbt-snowflake>=1.7.0
